+++ CHANGELOG.md
@@ -4,7 +4,7 @@
  - large memory leak fixed for non product import (association, product group, attribute, categories, etc...)
 
 ## BC breaks
- - protected postWrite method not called anymore from BaseCon-nectorBundle\\Writer\\Doctrine\\Writer.
+ - protected postWrite method not called anymore from BaseConnectorBundle\\Writer\\Doctrine\\Writer.
  If you need it, override the write method, call the parent and add your code after.
  - constructor of Pim\Bundle\BaseConnectorBundle\Writer\Doctrine\Writer has changed
  - Pim\Bundle\TransformBundle\Cache\ProductCacheClearer has been renamed Pim\Bundle\TransformBundle\Cache\CacheClearer