+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 1.2.2 (2014-09-05)
+
+## Improvements
+ - CacheClearer splits into two services, one for Product and one for other entities
+
+## Bug fixes
+ - association import with MongoDB fixes
+
 # 1.2.1 (2014-09-03)
 
 ## Bug fixes